+++ book-shop/controller/cartsItemsController.js
@@ -1,6 +1,7 @@
 const conn = require('../mariadb');
 const jwt = require('jsonwebtoken');
 const dotenv = require('dotenv');
+const ensureAuthorization = require('../auth');
 const { StatusCodes } = require('http-status-codes');
 
 dotenv.config();
@@ -28,7 +29,7 @@ const getCartItems = (req, res) => {
 	const { selected } = req.body;
 	const authorization = ensureAuthorization(req);
 
-	if (authorization instanceof TokenExpiredError) {
+	if (authorization instanceof jwt.TokenExpiredError) {
 		return res.status(StatusCodes.UNAUTHORIZED).json({ message: 'Token expired. Please log in again.' });
 	} else if (authorization instanceof jwt.JsonWebTokenError) {
 		return res.status(StatusCodes.BAD_REQUEST).json({
@@ -38,8 +39,16 @@ const getCartItems = (req, res) => {
 		let query = `SELECT cartItems.id, book_id, title, summary, quantity, price 
 				FROM cartItems LEFT JOIN books 
 				ON cartItems.book_id = books.id
-				WHERE user_id =? AND cartItems.id IN (?)`;
-		conn.query(query, [authorization.id, selected], (err, results) => {
+				WHERE user_id =?`;
+
+		let values = [authorization.id];
+
+		if (selected) {
+			query += ` AND cartItems.id IN (?)`;
+			values.push(selected);
+		}
+
+		conn.query(query, values, (err, results) => {
 			if (err) {
 				console.log(err);
 				return res.status(StatusCodes.BAD_REQUEST).json({ message: 'Internal Server Error' });
@@ -50,40 +59,52 @@ const getCartItems = (req, res) => {
 };
 
 const removeCartItem = (req, res) => {
-	const { cartItemId } = req.params;
+	let authorization = ensureAuthorization(req, res);
 
-	let query = 'DELETE FROM cartItems WHERE id = ?';
+	if (authorization instanceof jwt.TokenExpiredError) {
+		return res.status(StatusCodes.Unauthorized).json({
+			message: 'Token expired. Please log in again.',
+		});
+	} else if (authorization instanceof jwt.JsonWebTokenError) {
+		return res.status(StatusCodes.BAD_REQUEST).json({
+			message: 'Invalid token. Please log in again.',
+		});
+	} else {
+		const { cartItemId } = req.params;
 
-	conn.query(query, id, (err, results) => {
-		if (err) {
-			console.log(err);
-			return res.status(StatusCodes.BAD_REQUEST).json({ message: 'Internal Server Error' });
-		}
+		let query = 'DELETE FROM cartItems WHERE id = ?';
 
-		if (results) {
-			return conn.query('SELECT * FROM likes', (err, results1) => {
-				if (err) {
-					console.log(err);
-					return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: 'Internal Server Error' });
-				}
-				return res.status(StatusCodes.OK).json(results1);
-			});
-		}
-	});
+		conn.query(query, id, (err, results) => {
+			if (err) {
+				console.log(err);
+				return res.status(StatusCodes.BAD_REQUEST).json({ message: 'Internal Server Error' });
+			}
+
+			if (results) {
+				return conn.query('SELECT * FROM likes', (err, results1) => {
+					if (err) {
+						console.log(err);
+						return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: 'Internal Server Error' });
+					}
+					return res.status(StatusCodes.OK).json(results1);
+				});
+			}
+		});
+	}
 };
 
-function ensureAuthorization(req, res) {
-	try {
-		let receivedJWT = req.headers['authorization'];
-		console.log(receivedJWT);
+// function ensureAuthorization(req, res) {
+// 	try {
+// 		let receivedJWT = req.headers['authorization'];
+// 		console.log(receivedJWT);
 
-		let decodedUser = jwt.verify(receivedJWT, process.env.PRIVATE_KEY);
-		console.log(decodedUser);
+// 		let decodedUser = jwt.verify(receivedJWT, process.env.PRIVATE_KEY);
+// 		console.log(decodedUser);
 
-		return decodedUser;
-	} catch (error) {
-		return;
-	}
-}
+// 		return decodedUser;
+// 	} catch (error) {
+// 		return;
+// 	}
+// }
 
 module.exports = { addToCart, getCartItems, removeCartItem };
